pyproject.toml
[project]
name = "BackendManutencao"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Anthony Loche", email = "anthonylocheifc@gmail.com"},
    {name = "Jonatas Peraza", email = "jonatassilvaperaza@gmail.com"}
    ]
dependencies = [
    "django>=5.0.7",
    "djangorestframework>=3.15.2",
    "django-cors-headers>=4.4.0",
    "django-localflavor>=4.0",
    "django-extensions>=3.2.3",
    "pydotplus>=2.0.2",
    "weasyprint>=62.3",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"


[tool.pdm]
distribution = true

[tool.pdm.scripts]
runserver = "pdm run python manage.py runserver"
makemigrations = "pdm run python manage.py makemigrations"
migrate = "pdm run python manage.py migrate"
createsuperuser = "pdm run python manage.py createsuperuser"
setup = "python modelagem.py"
model = "python manage.py graph_models -a -o models.png"
